Deploy on DigitalOcean instead of Github Pages
Since I'm changing domains from smeriwether.com to stephen.fyi, I need
to set up a 301 redirect from the old domain to the new domain. So, the
easiest way to do this is to run a small DO server with those redirect
rules. And if I'm going to do that I might as well host the website
there too.